pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Mauricio Pisabarro <mauripisabarro@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.2"
djangorestframework = "^3.13.1"
psycopg2 = "^2.9.3"
pydantic = "^1.9.0"
django-environ = "^0.8.1"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
